Remove duplicated function
