+++ packages/extension/src/utils/hexConverter.test.ts
@@ -1,4 +1,4 @@
-import { currencyToHex, hexToCurrency } from './hexConverter';
+import { currencyToHex } from './hexConverter';
 
 describe('currencyToHex function', () => {
   it('should convert a string into hex and pad it to 40 characters', () => {
@@ -7,31 +7,5 @@ describe('currencyToHex function', () => {
 
     expect(hex).toEqual('534F4C4F00000000000000000000000000000000');
     expect(hex.length).toEqual(40);
-  });
-});
-
-describe('hexToCurrency', () => {
-  it('should convert hex to currency', () => {
-    const hex = '534F4C4F00000000000000000000000000000000';
-
-    const currency = hexToCurrency(hex);
-
-    expect(currency).toEqual('SOLO');
-  });
-
-  it('should trim trailing zeros in the hex string', () => {
-    const hex = '534F4C4F';
-
-    const currency = hexToCurrency(hex);
-
-    expect(currency).toEqual('SOLO');
-  });
-
-  it('should return empty string for empty input', () => {
-    const hex = '';
-
-    const currency = hexToCurrency(hex);
-
-    expect(currency).toEqual('');
   });
 });
